+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Breaking change
 - dict_out: Print dictionary to a file, standard out, and/or an email instance either in expanded or flatten JSON format.
 
 ### Changed
-- rpt_mst_log, rpt_slv_log: Converted output to dictionary and returned results to calling function.
+- rpt_mst_log, rpt_slv_log, chk_mst_log, chk_slv_thr, chk_slv_err: Converted output to dictionary and returned results to calling function.
 - main, run_program, call_run_chk: Replaced the use of arg_parser (args_array) with gen_class.ArgParser class (args).
 - main, call_run_chk: Changed "-b" option to "-i" option to be inline with the other programs.
 - config/mysql_cfg.py.TEMPLATE: Added TLS version entry.